refactor: deploy bff and playground locally
