+++ now/constants.py
@@ -4,8 +4,8 @@ from docarray.typing import Image, Text, Video
 
 from now.utils import BetterEnum
 
-DOCKER_BFF_PLAYGROUND_TAG = '0.0.146-refactor-remove-webcam-2'
-NOW_PREPROCESSOR_VERSION = '0.0.121-refactor-remove-k8s-4'
+DOCKER_BFF_PLAYGROUND_TAG = '0.0.146-refactor-remove-k8s-5'
+NOW_PREPROCESSOR_VERSION = '0.0.121-refactor-remove-k8s-5'
 NOW_ELASTIC_INDEXER_VERSION = '0.0.142-fix-list-endpoint-4'
 NOW_AUTOCOMPLETE_VERSION = '0.0.8-refactor-bump-jina-version-to-latest-17'
