+++ tests/integration/conftest.py
@@ -2,15 +2,17 @@ import logging
 import os
 import tempfile
 import time
+from multiprocessing import Process
+
 import hubble
 import json
-from argparse import Namespace
 import pytest
 from pytest_mock import MockerFixture
 
+from deployment.bff.app.app import run_server
+from deployment.playground.playground import deploy_streamlit
 from now.utils import get_flow_id
 from now.deployment.deployment import terminate_wolf
-from now.cli import cli
 
 logging.basicConfig(level=logging.DEBUG)
 log = logging.getLogger(__name__)
@@ -71,3 +73,22 @@ def cleanup(dataset):
             f'Time taken to execute deployment with dataset `{dataset}`: {mins}m {secs}s'
         )
         print(50 * '#')
+
+
+@pytest.fixture()
+def start_bff():
+    p1 = Process(target=run_server, args=(8080,))
+    p1.daemon = True
+    p1.start()
+    yield
+    p1.terminate()
+
+
+@pytest.fixture()
+def start_playground():
+    p1 = Process(target=deploy_streamlit)
+    p1.daemon = True
+    p1.start()
+    yield
+    p1.terminate()
+
